pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-agent-system"
version = "0.1.0"
description = "A multi-agent system with MCP servers and Groq-powered agents"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "PyYAML>=6.0",
    "groq>=0.4.1",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools]
packages = ["mcp_servers", "sub_agents", "supervisor", "src"]

[tool.uv.sources]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100
